fix pointer arithmetic for very large matrices

+++ magmablas/ztranspose_inplace.cu
@@ -110,7 +110,7 @@ __global__ void ztranspose_inplace_even(
     ii *= NB;
     jj *= NB;
 
-    magmaDoubleComplex *A = matrix + ii+i + (jj+j)*lda;
+    magmaDoubleComplex *A = matrix + ii+i + (size_t)(jj+j)*(size_t)lda;
     if ( ii == jj ) {
         if ( ii+i < n && jj+j < n ) {
             sA[j][i] = *A;
@@ -121,7 +121,7 @@ __global__ void ztranspose_inplace_even(
         }
     }
     else {
-        magmaDoubleComplex *B = matrix + jj+i + (ii+j)*lda;
+        magmaDoubleComplex *B = matrix + jj+i + (size_t)(ii+j)*(size_t)lda;
         if ( ii+i < n && jj+j < n ) {
             sA[j][i] = *A;
         }
